requirements.txt
fastapi
starlette
python-multipart
uvicorn
pydantic
ldap3
pyjwt
pandas
jinjasql
aiofiles
openpyxl
xlrd
redis
numpy
requests
zeep
Jinja2
pdfkit
pyyaml
boto3
psycopg[binary]
pydantic[dotenv]
botocore
aiosmtplib
XlsxWriter
pydantic-settings
